+++ app/src/main/java/corp/kairos/adamastor/Onboarding/Onboard1WelcomeActivity.java
@@ -14,12 +14,7 @@ import corp.kairos.adamastor.R;
 import corp.kairos.adamastor.Settings.Settings;
 
 public class Onboard1WelcomeActivity extends AnimationCompatActivity {
-    private boolean permissionsGranted = false;
-    public static final int MY_PERMISSIONS_REQUEST_LOCATION = 99;
     private Settings settingsUser;
-
-    private static final String TAG = Onboard1WelcomeActivity.class.getName();
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -28,7 +23,6 @@ public class Onboard1WelcomeActivity extends AnimationCompatActivity {
         getSupportActionBar().setDisplayHomeAsUpEnabled(false);
         this.settingsUser = Settings.getInstance(this);
         this.settingsUser.resetSettings();
-        checkPermissions();
     }
     @Override
     protected void onResume() {
@@ -44,40 +38,6 @@ public class Onboard1WelcomeActivity extends AnimationCompatActivity {
     public void start(View v) {
         Intent i = new Intent(this,Onboard2SpecialPermissionActivity.class);
         startActivity(i);
-    }
-
-    private void checkPermissions(){
-        if ((ActivityCompat.checkSelfPermission(this,
-                android.Manifest.permission.ACCESS_FINE_LOCATION)
-                != PackageManager.PERMISSION_GRANTED) ||
-                (ActivityCompat.checkSelfPermission(this,
-                        android.Manifest.permission.ACCESS_COARSE_LOCATION)
-                        != PackageManager.PERMISSION_GRANTED)||
-                (ActivityCompat.checkSelfPermission(this,
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                        != PackageManager.PERMISSION_GRANTED) ||
-                (ActivityCompat.checkSelfPermission(this,
-                        Manifest.permission.GET_ACCOUNTS)
-                        != PackageManager.PERMISSION_GRANTED)){
-
-
-            requestPremissions();
-
-        }else{
-            Log.i(TAG, "permissions OK");
-        }
-    }
-
-    //KAIROS
-    private void requestPremissions(){
-        ActivityCompat.requestPermissions(this,
-                new String[]{Manifest.permission.ACCESS_COARSE_LOCATION,
-                        Manifest.permission.ACCESS_FINE_LOCATION,
-                        Manifest.permission.ACCOUNT_MANAGER,
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                        Manifest.permission.PACKAGE_USAGE_STATS,
-                        Manifest.permission.GET_ACCOUNTS},
-                MY_PERMISSIONS_REQUEST_LOCATION);
     }
 
 }
